Extract playback loop control
